bt-458: * move to folder validation
