+++ frontend/src/bundles/talent-onboarding/helpers/helpers.ts
@@ -12,5 +12,5 @@ export {
     toUrlLinks,
 } from './map-project-links/map-project-links.js';
 export { setEnglishLevelValue } from './set-english-level-value/set-english-level-value.js';
-export { validateFileSize } from './validate-file-size.js';
+export { validateFileSize } from './validate-file-size/validate-file-size.js';
 export { validateFileType } from './validate-file-type/validate-file-type.js';
