+++ frontend/src/bundles/talent-onboarding/helpers/validate-file-size/validate-file-size.ts
@@ -5,7 +5,7 @@ import {
     type UseFormSetError,
 } from 'react-hook-form';
 
-import { MAX_FILE_SIZE } from '../components/contacts-cv-step/constants/constants.js';
+import { MAX_FILE_SIZE } from '../../components/contacts-cv-step/constants/constants.js';
 
 const validateFileSize = <T extends FieldValues>({
     name,
